倍数 小数点隐藏

+++ 彩小蜜/Classes/Section/Home/FootballMatch/Model/FootballBetInfoModel.swift
@@ -16,7 +16,7 @@ class FootballBetInfoModel : HandyJSON{
     
     var maxLotteryMoney: Double!
     var betCells: [BetCellModel]!
-    var betNum: Double!
+    var betNum: Int!
     var betType: String!
     var lotteryPrints: [LotteryPrintsModel]!
     var maxBonus: String!
